Circumvent isolated world environments
Content scripts are isolated from the web page that they are executed
in. As a result, the google API loaded by the page could not be accessed
by the content_script.js file. To allow this file to do its work, inject
it into the page as as script.

See: https://stackoverflow.com/questions/9515704/
  insert-code-into-the-page-context-using-a-content-script